# api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Nirvana Backend API",
    "version": "1.0.0",
    "description": "Concise OpenAPI specification for frontend integration with the Nirvana backend (local dev at http://localhost:8080).\n\nAuthentication: JWT (Bearer). Register/login return a raw JWT string. Role claims: PATIENT or DOCTOR. Many endpoints require role-based access as noted."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "PATIENT",
              "DOCTOR"
            ]
          }
        },
        "required": [
          "username",
          "email",
          "password",
          "role"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ChatbotRequest": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ]
      },
      "ChatbotResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "sentAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MoodRequest": {
        "type": "object",
        "properties": {
          "MoodLabel": {
            "type": "string"
          },
          "MoodScore": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "Note": {
            "type": "string"
          },
          "LoggedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "MoodScore"
        ]
      },
      "MoodResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "moodScore": {
            "type": "integer"
          },
          "moodLabel": {
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "loggedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JournalRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "content"
        ]
      },
      "JournalResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "InviteRequest": {
        "type": "object",
        "properties": {
          "doctorEmail": {
            "type": "string",
            "format": "email"
          }
        },
        "required": [
          "doctorEmail"
        ]
      },
      "PatientSummary": {
        "type": "object",
        "properties": {
          "ConsentId": {
            "type": "integer"
          },
          "patientId": {
            "type": "integer"
          },
          "patientName": {
            "type": "string"
          },
          "patientEmail": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "DoctorSummary": {
        "type": "object",
        "properties": {
          "ConsentId": {
            "type": "integer"
          },
          "doctorId": {
            "type": "integer"
          },
          "doctorName": {
            "type": "string"
          },
          "doctorEmail": {
            "type": "string",
            "format": "email"
          }
        }
      },
      "WellnessResponse": {
        "type": "object",
        "properties": {
          "avgMoodScore": {
            "type": "number"
          },
          "avgFrequentMood": {
            "type": "string"
          },
          "totalMoodLogs": {
            "type": "integer"
          },
          "bestday": {
            "type": "string"
          },
          "worstday": {
            "type": "string"
          },
          "totalJournalEntries": {
            "type": "integer"
          },
          "aiInsight": {
            "type": "string"
          },
          "emotions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register new user (returns JWT)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token (plain text)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login (returns JWT)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JWT token (plain text)",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chatbot": {
      "post": {
        "summary": "Send chat message to AI chatbot (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatbotRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ChatbotResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatbotResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/chatbot/history": {
      "get": {
        "summary": "Get chat history for authenticated patient (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of ChatbotResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatbotResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mood": {
      "post": {
        "summary": "Log mood (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoodRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Mood created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MoodResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/mood/history": {
      "get": {
        "summary": "Get mood history for authenticated patient (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of MoodResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoodResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/mood/range": {
      "get": {
        "summary": "Get mood entries in a date range (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "ISO date-time start"
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            },
            "description": "ISO date-time end"
          }
        ],
        "responses": {
          "200": {
            "description": "Array of MoodResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoodResponse"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/mood/{id}": {
      "delete": {
        "summary": "Delete mood entry by id (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion message",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/mood/doctor/{patientId}/history": {
      "get": {
        "summary": "Doctor: get a patient's mood history (DOCTOR)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of MoodResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MoodResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (no consent)"
          }
        }
      }
    },
    "/journal/journal": {
      "post": {
        "summary": "Add a journal entry (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JournalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Journal created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JournalResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/journal/journal/history": {
      "get": {
        "summary": "Get journal history for authenticated patient (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of JournalResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/journal/doctor/{patientId}/history": {
      "get": {
        "summary": "Doctor: get a patient's journal history (DOCTOR)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "patientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of JournalResponse",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JournalResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden (no consent)"
          }
        }
      }
    },
    "/consent/invite": {
      "post": {
        "summary": "Patient invites a doctor by email (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plain text result",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/consent/{consentId}/revoke": {
      "post": {
        "summary": "Patient revokes a consent (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "consentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plain text result",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/consent/doctor/my-patients": {
      "get": {
        "summary": "Doctor: list patients who granted consent (DOCTOR)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of PatientSummary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PatientSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/consent/patient/my-doctors": {
      "get": {
        "summary": "Patient: list doctors the patient granted consent to (PATIENT)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Array of DoctorSummary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DoctorSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/wellness/report": {
      "get": {
        "summary": "Get wellness report (authenticated)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "WellnessResponse",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WellnessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "auth",
      "description": "Authentication endpoints"
    },
    {
      "name": "mood",
      "description": "Mood tracking"
    },
    {
      "name": "journal",
      "description": "Journal entries"
    },
    {
      "name": "chatbot",
      "description": "AI chatbot"
    },
    {
      "name": "consent",
      "description": "Consent management"
    },
    {
      "name": "wellness",
      "description": "Aggregated wellness reports"
    }
  ]
}